+++ spec/field_struct/types/uuid_spec.rb
@@ -40,7 +40,7 @@ RSpec.describe FieldStruct::Types::UUID do
     end
 
     it 'auto-populates the field options with default_format' do
-      expect(klass.metadata[:id].options[:format]).to eq(described_class::DEFAULT_FORMAT)
+      expect(klass.metadata[:id].options[:format]).to eq(described_class.default_format)
     end
 
     it 'accepts a valid UUID' do